fix: add code improvements based on static analysis

+++ RSSchool_T1/RSSchool_T1/Exercises/1/T1Array.m
@@ -29,7 +29,7 @@
             }
         }
         
-        for(int i = indiciesToRemove.count - 1; i >=0; i--)
+        for(NSInteger i = indiciesToRemove.count - 1; i >=0; i--)
         {
             [mutableSadArray removeObjectAtIndex:(((NSNumber *)indiciesToRemove[i]).intValue)];
         }
